fix(graphql): harden semantic retry evaluation
