+++ packages/graphql/src/services/stacks.ts
@@ -1,26 +1,9 @@
-import { createHash } from 'node:crypto'
-import {
-  computeAwardedXp,
-  computeSimpleAwardedPoints,
-  gradeQuestionCaseStudy,
-  gradeQuestionFreeText,
-  gradeQuestionKPRIM,
-  gradeQuestionMC,
-  gradeQuestionNumerical,
-  gradeQuestionSC,
-  gradeQuestionSelection,
-} from '@klicker-uzh/grading'
 import * as DB from '@klicker-uzh/prisma/client'
 import type {
   CaseStudyElementData,
-  CaseStudySolutionsObject,
-  Choice,
   ChoicesElementData,
   ChoicesResponse,
   ContentElementData,
-  ElementData,
-  ElementInstanceResults,
-  ElementOptionsAnswerCollection,
   ElementOptionsCaseStudy,
   ElementOptionsChoices,
   ElementOptionsFreeText,
@@ -35,15 +18,8 @@ import type {
   FlashcardElementData,
   FreeTextElementData,
   InstanceEvaluation,
-  InstanceEvaluationCaseStudy,
-  InstanceEvaluationChoices,
-  InstanceEvaluationFreeText,
-  InstanceEvaluationNumerical,
-  InstanceEvaluationSelection,
   NumericalElementData,
   SelectionElementData,
-  SingleCaseStudyResponse,
-  SingleQuestionResponse,
   SingleQuestionResponseCaseStudy,
   SingleQuestionResponseChoices,
   SingleQuestionResponseContent,
@@ -52,27 +28,56 @@ import type {
   SingleQuestionResponseValue,
 } from '@klicker-uzh/types'
 import { FlashcardCorrectness, StackFeedbackStatus } from '@klicker-uzh/types'
-import {
-  getInitialInstanceResults,
-  type PrismaTransactionClient,
-} from '@klicker-uzh/util'
+import type { PrismaTransactionClient } from '@klicker-uzh/util'
 import dayjs from 'dayjs'
-import { max, mean, median, min, quantileSeq, round, std } from 'mathjs'
-import { toLowerCase } from 'remeda'
-import type { ICaseStudyElementEvaluationResults } from '@/schema/evaluation.js'
+import { round } from 'mathjs'
 import type { Context, ContextWithUser } from '../lib/context.js'
-import type {
-  CaseStudyCaseResponse,
-  CaseStudyElementOptions,
-  ResponseInput,
-} from '../ops.js'
+import type { ResponseInput } from '../ops.js'
+import { submitSemanticFreeTextPracticeResponse } from './freeTextPracticeSubmission.js'
 import {
-  type CreateFreeTextAttemptInput,
-  createFreeTextAttempt,
-  type FreeTextEvaluationServiceOptions,
-} from './freeTextEvaluation.js'
-import { applyFreeTextAttemptResponseInTransaction as applyFreeTextAttemptResponseCore } from './freeTextResponseApplication.js'
-import { upsertDailyTimelineEntry } from './participants.js'
+  combineCaseStudyResults,
+  combineChoicesResults,
+  combineFreeTextResults,
+  combineNumericalResults,
+  combineSelectionResults,
+  computeNumericalStatistics,
+  evaluateCaseStudyElementResponse,
+  evaluateChoicesElementResponse,
+  evaluateFreeTextElementResponse,
+  evaluateNumericalElementResponse,
+  evaluateSelectionElementResponse,
+} from './questionResponseEvaluation.js'
+import { applyQuestionResponseInTransaction } from './questionResponsePersistence.js'
+import {
+  evaluateCaseStudyAnswerCorrectness,
+  evaluateChoicesAnswerCorrectness,
+  evaluateFreeTextAnswerCorrectness,
+  evaluateNumericalAnswerCorrectness,
+  evaluateSelectionAnswerCorrectness,
+} from './questionResponseResults.js'
+import {
+  combineCorrectnessParams,
+  combineNewCorrectnessParams,
+  computeNewAverageTimes,
+  computeUpdatedInstanceStatistics,
+  type SpacedRepetitionResult,
+  updateSpacedRepetition,
+} from './responseTracking.js'
+
+export { POINTS_PER_INSTANCE } from './questionResponseEvaluation.js'
+export {
+  evaluateCaseStudyAnswerCorrectness,
+  evaluateChoicesAnswerCorrectness,
+  evaluateFreeTextAnswerCorrectness,
+  evaluateNumericalAnswerCorrectness,
+  evaluateSelectionAnswerCorrectness,
+  updateCaseStudyResults,
+  updateChoicesResults,
+  updateFreeTextResults,
+  updateNumericalResults,
+  updateSelectionResults,
+} from './questionResponseResults.js'
+export { updateSpacedRepetition } from './responseTracking.js'
 
 type ExistingInstanceType = DB.ElementInstance & {
   elementStack?: {
@@ -80,10 +85,6 @@ type ExistingInstanceType = DB.ElementInstance & {
     microLearningId?: string | null
   } | null
 }
-
-export const POINTS_PER_INSTANCE = 10
-const POINTS_AWARD_TIMEFRAME_DAYS = 6
-const XP_AWARD_TIMEFRAME_DAYS = 1
 
 const flashcardResultMap: Record<FlashcardCorrectness, StackFeedbackStatus> = {
   [FlashcardCorrectness.INCORRECT]: StackFeedbackStatus.INCORRECT,
@@ -93,251 +94,6 @@ const flashcardResultMap: Record<FlashcardCorrectness, StackFeedbackStatus> = {
 
 // ! Shared Helper Functions
 // #region
-interface CombineCorrectnessParamsInput {
-  correct: boolean
-  partial: boolean
-  incorrect: boolean
-  existingResponse?: DB.QuestionResponse | null
-}
-
-function combineNewCorrectnessParams({
-  correct,
-  partial,
-  incorrect,
-}: CombineCorrectnessParamsInput) {
-  return {
-    // track last answer date
-    lastAnsweredAt: new Date(),
-
-    // CORRECT
-    correctCount: correct ? 1 : 0,
-    correctCountStreak: correct ? 1 : 0,
-    lastCorrectAt: correct ? new Date() : undefined,
-
-    // PARTIALLY CORRECT
-    partialCorrectCount: partial ? 1 : 0,
-    lastPartialCorrectAt: partial ? new Date() : undefined,
-
-    // WRONG
-    wrongCount: incorrect ? 1 : 0,
-    lastWrongAt: incorrect ? new Date() : undefined,
-  }
-}
-
-function combineCorrectnessParams({
-  correct,
-  partial,
-  incorrect,
-  existingResponse,
-}: CombineCorrectnessParamsInput) {
-  return {
-    // track last answer date
-    lastAnsweredAt: new Date(),
-
-    // CORRECT
-    correctCount: {
-      increment: correct ? 1 : 0,
-    },
-    correctCountStreak: {
-      increment: correct
-        ? 1
-        : existingResponse
-          ? -existingResponse.correctCountStreak
-          : 0,
-    },
-    lastCorrectAt: correct ? new Date() : undefined,
-
-    // PARTIALLY CORRECT
-    partialCorrectCount: {
-      increment: partial ? 1 : 0,
-    },
-    lastPartialCorrectAt: partial ? new Date() : undefined,
-
-    // INCORRECT
-    wrongCount: {
-      increment: incorrect ? 1 : 0,
-    },
-    lastWrongAt: incorrect ? new Date() : undefined,
-  }
-}
-
-type SpacedRepetitionResult = {
-  efactor: number
-  interval: number
-  nextDueAt: Date
-}
-
-export function updateSpacedRepetition({
-  eFactor,
-  interval,
-  streak,
-  grade,
-}: {
-  eFactor: number
-  interval: number
-  streak: number
-  grade: number
-}): SpacedRepetitionResult {
-  if (grade < 0 || grade > 1) {
-    throw new Error('Grade must be between 0 and 1.')
-  }
-
-  // scale grade to 0-5 range (definition of algorithm)
-  const scaledGrade = grade * 5
-
-  // update efactor and interval
-  let newEfactor = Math.max(
-    1.3,
-    eFactor + (0.1 - (5 - scaledGrade) * (0.08 + (5 - scaledGrade) * 0.02))
-  )
-  newEfactor = parseFloat(newEfactor.toFixed(2))
-
-  let newInterval: number
-  if (scaledGrade < 3) {
-    newInterval = 1
-  } else {
-    if (streak === 1) {
-      newInterval = 2
-    } else if (streak === 2) {
-      newInterval = 6
-    } else {
-      newInterval = Math.ceil(interval * newEfactor)
-    }
-  }
-
-  // ensure that maximum interval is bounded
-  newInterval = Math.min(newInterval, 36500)
-
-  // compute next due date to sort by (=> spaced repetition)
-  const nextDueAt = dayjs().add(newInterval, 'day').toDate()
-
-  return {
-    efactor: newEfactor,
-    interval: newInterval,
-    nextDueAt: nextDueAt,
-  }
-}
-
-function computeNewAverageTimes({
-  existingInstance,
-  existingResponse,
-  answerTime,
-}: {
-  existingInstance: DB.ElementInstance & {
-    instanceStatistics: DB.InstanceStatistics | null
-  }
-  existingResponse: DB.QuestionResponse | null
-  answerTime: number
-}): { newAverageResponseTime: number; newAverageInstanceTime: number } {
-  const existingParticipantCount =
-    existingInstance.instanceStatistics!.uniqueParticipantCount
-  const existingInstanceTime =
-    existingInstance.instanceStatistics!.averageTimeSpent
-  const newAverageResponseTime = existingResponse
-    ? (existingResponse.averageTimeSpent * existingResponse.trialsCount +
-        answerTime) /
-      (existingResponse.trialsCount + 1)
-    : answerTime
-  const newAverageInstanceTime = existingResponse
-    ? (existingInstanceTime! * existingParticipantCount -
-        existingResponse.averageTimeSpent +
-        answerTime) /
-      existingParticipantCount
-    : ((existingInstanceTime ?? 0) * existingParticipantCount + answerTime) /
-      (existingParticipantCount + 1)
-
-  return { newAverageResponseTime, newAverageInstanceTime }
-}
-
-function computeUpdatedInstanceStatistics({
-  participation,
-  existingResponse,
-  newAverageInstanceTime,
-  answerCorrect,
-  answerPartial,
-  answerIncorrect,
-  instanceInPracticeQuiz,
-}: {
-  participation: DB.Participation | null
-  existingResponse: DB.QuestionResponse | null
-  newAverageInstanceTime?: number
-  answerCorrect: boolean
-  answerPartial: boolean
-  answerIncorrect: boolean
-  instanceInPracticeQuiz: boolean
-}) {
-  return participation
-    ? {
-        update: {
-          uniqueParticipantCount: {
-            increment: Number(!existingResponse),
-          },
-          averageTimeSpent: newAverageInstanceTime ?? 0,
-          correctCount: {
-            increment: Number(answerCorrect),
-          },
-          partialCorrectCount: {
-            increment: Number(answerPartial),
-          },
-          wrongCount: {
-            increment: Number(answerIncorrect),
-          },
-          firstCorrectCount: {
-            increment: Number(
-              answerCorrect && !existingResponse && instanceInPracticeQuiz
-            ),
-          },
-          firstPartialCorrectCount: {
-            increment: Number(
-              answerPartial && !existingResponse && instanceInPracticeQuiz
-            ),
-          },
-          firstWrongCount: {
-            increment: Number(
-              answerIncorrect && !existingResponse && instanceInPracticeQuiz
-            ),
-          },
-          lastCorrectCount: {
-            increment:
-              Number(answerCorrect && instanceInPracticeQuiz) -
-              Number(
-                existingResponse?.lastResponseCorrectness ===
-                  DB.ResponseCorrectness.CORRECT
-              ),
-          },
-          lastPartialCorrectCount: {
-            increment:
-              Number(answerPartial && instanceInPracticeQuiz) -
-              Number(
-                existingResponse?.lastResponseCorrectness ===
-                  DB.ResponseCorrectness.PARTIAL
-              ),
-          },
-          lastWrongCount: {
-            increment:
-              Number(answerIncorrect && instanceInPracticeQuiz) -
-              Number(
-                existingResponse?.lastResponseCorrectness ===
-                  DB.ResponseCorrectness.WRONG
-              ),
-          },
-        },
-      }
-    : {
-        update: {
-          anonymousCorrectCount: {
-            increment: Number(answerCorrect),
-          },
-          anonymousPartialCorrectCount: {
-            increment: Number(answerPartial),
-          },
-          anonymousWrongCount: {
-            increment: Number(answerIncorrect),
-          },
-        },
-      }
-}
-
 export function combineStackStatus({
   prevStatus,
   newStatus,
@@ -1159,1456 +915,6 @@ async function respondToContent(
 
 // ! Question Response Logic
 // #region
-type SharedEvaluationProps =
-  | 'elementType'
-  | 'feedbacks'
-  | 'numAnswers'
-  | 'score'
-  | 'xp'
-  | 'percentile'
-  | 'pointsMultiplier'
-  | 'explanation'
-
-type ChoicesEvaluationReturnType = Pick<
-  InstanceEvaluationChoices,
-  SharedEvaluationProps | 'choices'
->
-type NumericalEvaluationReturnType = Pick<
-  InstanceEvaluationNumerical,
-  SharedEvaluationProps | 'solutionRanges' | 'exactSolutions' | 'responses'
->
-type FreeTextEvaluationReturnType = Pick<
-  InstanceEvaluationFreeText,
-  SharedEvaluationProps | 'solutions' | 'answers'
->
-type SelectionEvaluationReturnType = Pick<
-  InstanceEvaluationSelection,
-  SharedEvaluationProps | 'answerSolutionIds' | 'selectionResponses'
->
-type CaseStudyEvaluationReturnType = Pick<
-  InstanceEvaluationCaseStudy,
-  SharedEvaluationProps | 'assessments' | 'studySolutions'
->
-
-async function getValidateElementInstance({
-  prisma,
-  id,
-  participantId,
-}: {
-  prisma: PrismaTransactionClient
-  id: number
-  participantId?: string
-}) {
-  const existingInstance = await prisma.elementInstance.findUnique({
-    where: {
-      id,
-    },
-    include: {
-      elementStack: true,
-      instanceStatistics: true,
-      responses: participantId
-        ? {
-            where: {
-              participantId,
-            },
-          }
-        : false,
-    },
-  })
-
-  return existingInstance
-}
-
-function evaluateChoicesElementResponse({
-  elementData,
-  results,
-  anonymousResults,
-  correctness,
-  multiplier,
-}: {
-  elementData: ChoicesElementData
-  results: ElementResultsChoices
-  anonymousResults: ElementResultsChoices
-  correctness: number | null
-  multiplier?: number
-}): ChoicesEvaluationReturnType | null {
-  return {
-    elementType: elementData.type,
-    feedbacks: elementData.options.choices,
-    numAnswers: results.total + anonymousResults.total,
-    choices: combineChoicesResults({
-      choices: elementData.options.choices,
-      results: results.choices,
-      anonymousResults: anonymousResults.choices,
-    }),
-    score: computeSimpleAwardedPoints({
-      points: POINTS_PER_INSTANCE,
-      pointsPercentage: correctness,
-      pointsMultiplier: multiplier,
-    }),
-    xp: computeAwardedXp({
-      pointsPercentage: correctness,
-    }),
-    percentile: correctness ?? 0,
-    pointsMultiplier: multiplier ?? 1,
-    explanation: elementData.explanation,
-  }
-}
-
-function evaluateNumericalElementResponse({
-  elementData,
-  results,
-  anonymousResults,
-  correctness,
-  multiplier,
-}: {
-  elementData: NumericalElementData
-  results: ElementResultsOpen
-  anonymousResults: ElementResultsOpen
-  correctness: number | null
-  multiplier?: number
-}): NumericalEvaluationReturnType | null {
-  return {
-    elementType: DB.ElementType.NUMERICAL,
-    feedbacks: [],
-    numAnswers: results.total + anonymousResults.total,
-    responses: combineNumericalResults({
-      results,
-      anonymousResults,
-    }),
-    score: correctness
-      ? correctness * POINTS_PER_INSTANCE * (multiplier ?? 1)
-      : 0,
-    xp: computeAwardedXp({
-      pointsPercentage: correctness,
-    }),
-    percentile: correctness ?? 0,
-    pointsMultiplier: multiplier ?? 1,
-    explanation: elementData.explanation,
-    solutionRanges: elementData.options.solutionRanges ?? [],
-    exactSolutions: elementData.options.exactSolutions ?? [],
-  }
-}
-
-function evaluateFreeTextElementResponse({
-  elementData,
-  results,
-  anonymousResults,
-  correctness,
-  multiplier,
-}: {
-  elementData: FreeTextElementData
-  results: ElementResultsOpen
-  anonymousResults: ElementResultsOpen
-  correctness: number | null
-  multiplier?: number
-}): FreeTextEvaluationReturnType | null {
-  return {
-    elementType: DB.ElementType.FREE_TEXT,
-    feedbacks: [],
-    numAnswers: results.total + anonymousResults.total,
-    answers: combineFreeTextResults({
-      results,
-      anonymousResults,
-    }),
-    score: correctness
-      ? correctness * POINTS_PER_INSTANCE * (multiplier ?? 1)
-      : 0,
-    xp: computeAwardedXp({
-      pointsPercentage: correctness,
-    }),
-    percentile: correctness ?? 0,
-    pointsMultiplier: multiplier ?? 1,
-    explanation: elementData.explanation,
-    solutions: elementData.options.solutions ?? [],
-  }
-}
-
-function evaluateSelectionElementResponse({
-  elementData,
-  results,
-  anonymousResults,
-  correctness,
-  multiplier,
-}: {
-  elementData: SelectionElementData
-  results: ElementResultsSelection
-  anonymousResults: ElementResultsSelection
-  correctness: number | null
-  multiplier?: number
-}): SelectionEvaluationReturnType | null {
-  return {
-    elementType: DB.ElementType.SELECTION,
-    feedbacks: [],
-    numAnswers: results.total + anonymousResults.total,
-    selectionResponses: combineSelectionResults({
-      results,
-      anonymousResults,
-      answerOptions: elementData.options.answerCollection!,
-    }),
-    score: correctness
-      ? Math.round(correctness * POINTS_PER_INSTANCE * (multiplier ?? 1))
-      : 0,
-    xp: computeAwardedXp({
-      pointsPercentage: correctness,
-    }),
-    percentile: correctness ?? 0,
-    pointsMultiplier: multiplier ?? 1,
-    explanation: elementData.explanation,
-    answerSolutionIds: elementData.options.answerCollectionSolutionIds ?? [],
-  }
-}
-
-function evaluateCaseStudyElementResponse({
-  elementData,
-  results,
-  anonymousResults,
-  correctness,
-  multiplier,
-}: {
-  elementData: CaseStudyElementData
-  results: ElementResultsCaseStudy
-  anonymousResults: ElementResultsCaseStudy
-  correctness: number | null
-  multiplier?: number
-}): CaseStudyEvaluationReturnType | null {
-  return {
-    elementType: DB.ElementType.CASE_STUDY,
-    feedbacks: [],
-    numAnswers: results.total + anonymousResults.total,
-    assessments: reduceCaseStudyResults({
-      results,
-      anonymousResults,
-      options: elementData.options,
-    }),
-    studySolutions: elementData.options.cases.map((c) => ({
-      caseId: c.id,
-      solutions: elementData.options.hasSampleSolution ? c.solutions! : [],
-    })),
-    score: correctness
-      ? Math.round(correctness * POINTS_PER_INSTANCE * (multiplier ?? 1))
-      : 0,
-    xp: computeAwardedXp({
-      pointsPercentage: correctness,
-    }),
-    percentile: correctness ?? 0,
-    pointsMultiplier: multiplier ?? 1,
-    explanation: elementData.explanation,
-  }
-}
-
-function computeQuestionEvaluation({
-  elementData,
-  results,
-  anonymousResults,
-  correctness,
-  multiplier,
-}: {
-  elementData: ElementData
-  results: ElementInstanceResults
-  anonymousResults: ElementInstanceResults
-  correctness: number | null
-  multiplier?: number
-}) {
-  if (
-    (elementData.type === DB.ElementType.SC ||
-      elementData.type === DB.ElementType.MC ||
-      elementData.type === DB.ElementType.KPRIM) &&
-    'choices' in results &&
-    'choices' in anonymousResults
-  ) {
-    return evaluateChoicesElementResponse({
-      elementData,
-      results,
-      anonymousResults,
-      correctness,
-      multiplier,
-    })
-  } else if (
-    elementData.type === DB.ElementType.NUMERICAL &&
-    'responses' in results &&
-    'responses' in anonymousResults
-  ) {
-    return evaluateNumericalElementResponse({
-      elementData,
-      results,
-      anonymousResults,
-      correctness,
-      multiplier,
-    })
-  } else if (
-    elementData.type === DB.ElementType.FREE_TEXT &&
-    'responses' in results &&
-    'responses' in anonymousResults
-  ) {
-    return evaluateFreeTextElementResponse({
-      elementData,
-      results,
-      anonymousResults,
-      correctness,
-      multiplier,
-    })
-  } else if (
-    elementData.type === DB.ElementType.SELECTION &&
-    'selections' in results &&
-    'selections' in anonymousResults
-  ) {
-    return evaluateSelectionElementResponse({
-      elementData,
-      results,
-      anonymousResults,
-      correctness,
-      multiplier,
-    })
-  } else if (
-    elementData.type === DB.ElementType.CASE_STUDY &&
-    'assessments' in results &&
-    'assessments' in anonymousResults
-  ) {
-    return evaluateCaseStudyElementResponse({
-      elementData,
-      results,
-      anonymousResults,
-      correctness,
-      multiplier,
-    })
-  } else {
-    return null
-  }
-}
-
-export function evaluateChoicesAnswerCorrectness({
-  elementData,
-  response,
-}: {
-  elementData: ChoicesElementData
-  response: ResponseInput
-}) {
-  if (
-    !('choices' in response) ||
-    response.choices === null ||
-    typeof response.choices === 'undefined' ||
-    ((elementData.type === DB.ElementType.SC ||
-      elementData.type === DB.ElementType.MC) &&
-      response.choices.length === 0)
-  ) {
-    return null
-  }
-
-  const elementOptions = elementData.options
-  const solution = elementOptions.choices.reduce<number[]>((acc, choice) => {
-    if (choice.correct) return [...acc, choice.ix]
-    return acc
-  }, [])
-
-  if (elementData.type === DB.ElementType.SC) {
-    const correctness = gradeQuestionSC({
-      responseCount: elementOptions.choices.length,
-      response: response.choices,
-      solution,
-    })
-    return correctness
-  } else if (elementData.type === DB.ElementType.MC) {
-    const correctness = gradeQuestionMC({
-      responseCount: elementOptions.choices.length,
-      response: response.choices,
-      solution,
-    })
-    return correctness
-  } else {
-    const correctness = gradeQuestionKPRIM({
-      responseCount: elementOptions.choices.length,
-      response: response.choices,
-      solution,
-    })
-    return correctness
-  }
-}
-
-export function evaluateNumericalAnswerCorrectness({
-  elementData,
-  response,
-}: {
-  elementData: NumericalElementData
-  response: ResponseInput
-}) {
-  if (
-    !('value' in response) ||
-    response.value === null ||
-    typeof response.value === 'undefined'
-  ) {
-    return null
-  }
-
-  const correctness = gradeQuestionNumerical({
-    response: parseFloat(String(response.value)),
-    solutionRanges: elementData.options.solutionRanges ?? [],
-    exactSolutions: elementData.options.exactSolutions ?? [],
-  })
-  return correctness
-}
-
-export function evaluateFreeTextAnswerCorrectness({
-  elementData,
-  response,
-  treatFTDefaultCorrect = false,
-}: {
-  elementData: FreeTextElementData
-  response: ResponseInput
-  treatFTDefaultCorrect?: boolean
-}) {
-  // if the corresponding option is activated, treat FT questions without a sample solution always as correct
-  if (treatFTDefaultCorrect && !elementData.options.hasSampleSolution) {
-    return 1
-  }
-
-  if (
-    !('value' in response) ||
-    response.value === null ||
-    typeof response.value === 'undefined'
-  ) {
-    return null
-  }
-
-  const correctness = gradeQuestionFreeText({
-    response: response.value,
-    solutions: elementData.options.solutions ?? [],
-  })
-  return correctness
-}
-
-export function evaluateSelectionAnswerCorrectness({
-  elementData,
-  response,
-}: {
-  elementData: SelectionElementData
-  response: ResponseInput
-}) {
-  if (!elementData.options.hasSampleSolution) {
-    return 1
-  }
-
-  if (
-    !('selection' in response) ||
-    !response.selection ||
-    response.selection.length === 0
-  ) {
-    return null
-  }
-
-  const correctness = gradeQuestionSelection({
-    numberOfInputs: elementData.options.numberOfInputs!,
-    response: response.selection,
-    correctAnswers: elementData.options.answerCollectionSolutionIds,
-  })
-  return correctness
-}
-
-export function evaluateCaseStudyAnswerCorrectness({
-  elementData,
-  response,
-}: {
-  elementData: CaseStudyElementData
-  response: ResponseInput
-}) {
-  if (!elementData.options.hasSampleSolution) {
-    return 1
-  }
-
-  if (
-    !('assessment' in response) ||
-    !response.assessment ||
-    response.assessment.length === 0
-  ) {
-    return null
-  }
-
-  const hasSolutions = elementData.options.cases.every(
-    (caseItem) => caseItem.solutions && caseItem.solutions.length > 0
-  )
-  const correctness = hasSolutions
-    ? gradeQuestionCaseStudy({
-        response: response.assessment,
-        solutions: elementData.options.cases.map((caseItem) => ({
-          caseId: caseItem.id,
-          itemSolutions: caseItem.solutions!,
-        })),
-      })
-    : null
-  return correctness
-}
-
-export function updateChoicesResults({
-  previousResults,
-  response,
-}: {
-  previousResults: ElementResultsChoices
-  response: ResponseInput
-}): { results: ElementResultsChoices; modified: boolean } {
-  const results = previousResults
-  const updatedResults = results
-
-  if (
-    !('choices' in response) ||
-    response.choices === null ||
-    typeof response.choices === 'undefined'
-  ) {
-    return { results, modified: false }
-  }
-
-  updatedResults.choices = (
-    response as SingleQuestionResponseChoices
-  ).choices.reduce((acc, choiceResponse) => {
-    acc[choiceResponse.ix] = (acc[choiceResponse.ix] ?? 0) + 1
-    return acc
-  }, results.choices)
-  updatedResults.total = results.total + 1
-  return { results: updatedResults, modified: true }
-}
-
-export function updateNumericalResults({
-  previousResults,
-  elementData,
-  response,
-  correct,
-}: {
-  previousResults: ElementResultsOpen
-  elementData: ElementData
-  response: ResponseInput
-  correct?: boolean
-}): { results: ElementResultsOpen; modified: boolean } {
-  // verify the input types
-  if (elementData.type !== DB.ElementType.NUMERICAL) {
-    return { results: previousResults, modified: false }
-  }
-
-  const MD5 = createHash('md5')
-  const results = previousResults
-  const updatedResults = results
-
-  // validate format of response
-  if (
-    !('value' in response) ||
-    typeof response.value === 'undefined' ||
-    response.value === null ||
-    response.value === ''
-  ) {
-    return { results, modified: false }
-  }
-
-  // make sure that restrictions are fulfilled
-  const parsedValue = parseFloat(response.value)
-  if (
-    isNaN(parsedValue) ||
-    (typeof elementData.options.restrictions?.min === 'number' &&
-      parsedValue < elementData.options.restrictions.min) ||
-    (typeof elementData.options.restrictions?.max === 'number' &&
-      parsedValue > elementData.options.restrictions.max) ||
-    parsedValue > 1e30 || // prevent overflow
-    parsedValue < -1e30 // prevent underflow
-  ) {
-    return { results, modified: false }
-  }
-
-  const value = String(parsedValue)
-  MD5.update(value)
-  const hashedValue = MD5.digest('hex')
-
-  if (Object.keys(results.responses).includes(hashedValue)) {
-    updatedResults.responses = {
-      ...results.responses,
-      [hashedValue]: {
-        ...results.responses[hashedValue]!,
-        count: results.responses[hashedValue]!.count + 1,
-      },
-    }
-  } else {
-    updatedResults.responses = {
-      ...results.responses,
-      [hashedValue]: { value: value, count: 1, correct: correct },
-    }
-  }
-  updatedResults.total = results.total + 1
-  return { results: updatedResults, modified: true }
-}
-
-export function updateFreeTextResults({
-  previousResults,
-  elementData,
-  response,
-  correct,
-}: {
-  previousResults: ElementResultsOpen
-  elementData: ElementData
-  response: ResponseInput
-  correct?: boolean
-}): { results: ElementResultsOpen; modified: boolean } {
-  // verify the input types
-  if (elementData.type !== DB.ElementType.FREE_TEXT) {
-    return { results: previousResults, modified: false }
-  }
-
-  const MD5 = createHash('md5')
-  const results = previousResults
-  const updatedResults = results
-
-  // validate format of response and check that restrictions are fulfilled
-  if (
-    !('value' in response) ||
-    typeof response.value === 'undefined' ||
-    response.value === null ||
-    response.value === '' ||
-    (typeof elementData.options.restrictions?.maxLength === 'number' &&
-      response.value.length > elementData.options.restrictions?.maxLength)
-  ) {
-    return { results, modified: false }
-  }
-
-  const value = toLowerCase(response.value.trim())
-  MD5.update(value)
-  const hashedValue = MD5.digest('hex')
-
-  if (Object.keys(results.responses).includes(hashedValue)) {
-    updatedResults.responses = {
-      ...results.responses,
-      [hashedValue]: {
-        ...results.responses[hashedValue]!,
-        count: results.responses[hashedValue]!.count + 1,
-      },
-    }
-  } else {
-    updatedResults.responses = {
-      ...results.responses,
-      [hashedValue]: {
-        value: value,
-        count: 1,
-        correct: correct,
-      },
-    }
-  }
-  updatedResults.total = results.total + 1
-  return { results: updatedResults, modified: true }
-}
-
-export function updateSelectionResults({
-  previousResults,
-  response,
-}: {
-  previousResults: ElementResultsSelection
-  response: ResponseInput
-}) {
-  if (
-    !('selection' in response) ||
-    !response.selection ||
-    response.selection.length === 0
-  ) {
-    return { results: previousResults, modified: false }
-  }
-
-  // increment all values in updatedSelections that are contained in response.selection
-  const updatedSelections = { ...previousResults.selections }
-  response.selection.forEach((ix) => {
-    if (ix in updatedSelections && typeof updatedSelections[ix] === 'number') {
-      updatedSelections[ix] = updatedSelections[ix] + 1
-    }
-  })
-
-  return {
-    results: {
-      selections: updatedSelections,
-      total: previousResults.total + 1,
-    },
-    modified: true,
-  }
-}
-
-function convertCaseStudySolutionsObject({
-  instance,
-}: {
-  instance: DB.ElementInstance
-}): CaseStudySolutionsObject | undefined {
-  // convert case study solutions to object for faster access (if sample solution is defined)
-  const options = instance.elementData.options as CaseStudyElementOptions
-  const caseStudySolutions = options.hasSampleSolution
-    ? options.cases.reduce<CaseStudySolutionsObject>((acc, caseObj) => {
-        acc[caseObj.id] = caseObj.solutions!.reduce(
-          (itemAcc, { itemId, criteriaSolutions }) => {
-            itemAcc[String(itemId)] = criteriaSolutions.reduce(
-              (criterionAcc, { criterionId, min, max }) => {
-                criterionAcc[criterionId] = { min, max }
-                return criterionAcc
-              },
-              {}
-            )
-            return itemAcc
-          },
-          {}
-        )
-        return acc
-      }, {})
-    : undefined
-
-  return caseStudySolutions
-}
-
-export function updateCaseStudyResults({
-  previousResults,
-  response,
-  solutions,
-}: {
-  previousResults: ElementResultsCaseStudy
-  response: ResponseInput
-  solutions?: CaseStudySolutionsObject
-}) {
-  if (
-    !('assessment' in response) ||
-    !response.assessment ||
-    response.assessment.length === 0
-  ) {
-    return { results: previousResults, modified: false }
-  }
-
-  // update aggregated assessments wherever new values are provided
-  const newAssessments = { ...previousResults.assessments }
-  response.assessment.forEach((caseResponse) => {
-    const caseId = caseResponse.caseId
-
-    caseResponse.itemResponses.forEach((itemResponse) => {
-      const itemId = itemResponse.itemId
-
-      itemResponse.criterionResponses.forEach((criterionResponse) => {
-        const criterionId = criterionResponse.criterionId
-        const responseValue = criterionResponse.response
-
-        // hash response value for efficient access (as done for open results)
-        const MD5 = createHash('md5')
-        MD5.update(String(responseValue))
-        const responseHash = MD5.digest('hex')
-
-        // get existing responses for this case, item, and criterion
-        const existingCombinedResponses =
-          newAssessments[caseId]?.[String(itemId)]?.[criterionId]
-
-        // compute correctness of new response value
-        const sampleSolution =
-          solutions?.[caseId]?.[String(itemId)]?.[criterionId]
-        const responseCorrectness = sampleSolution
-          ? responseValue >= sampleSolution.min - Number.EPSILON &&
-            responseValue <= sampleSolution.max + Number.EPSILON
-          : undefined
-
-        // update existing response or create new response
-        if (!existingCombinedResponses) {
-          // even on initialization, all keys should already be set correctly
-          throw new Error('Existing combined responses are missing')
-        } else {
-          // increment counter of existing identical response or create new entry otherwise
-          if (Object.keys(existingCombinedResponses).includes(responseHash)) {
-            newAssessments[caseId]![String(itemId)]![criterionId]![
-              responseHash
-            ] = {
-              ...existingCombinedResponses[responseHash]!,
-              count: existingCombinedResponses[responseHash]!.count + 1,
-            }
-          } else {
-            newAssessments[caseId]![String(itemId)]![criterionId]![
-              responseHash
-            ] = {
-              value: responseValue,
-              count: 1,
-              correct: responseCorrectness,
-            }
-          }
-        }
-      })
-    })
-  })
-
-  return {
-    results: {
-      assessments: newAssessments,
-      total: previousResults.total + 1,
-    },
-    modified: true,
-  }
-}
-
-function updateQuestionResults({
-  existingInstance,
-  participation,
-  response,
-  caseStudySolutions,
-  correctnessOverride,
-}: {
-  existingInstance: DB.ElementInstance
-  participation: (DB.Participation & { participant: DB.Participant }) | null
-  response: ResponseInput
-  caseStudySolutions?: CaseStudySolutionsObject
-  correctnessOverride?: number
-}): {
-  correctness: number | null
-  results: ElementInstanceResults
-  modified: boolean
-} {
-  let correctness: number | null
-  const elementData = existingInstance.elementData
-  const previousResults = participation
-    ? existingInstance.results
-    : existingInstance.anonymousResults
-
-  if (
-    (elementData.type === DB.ElementType.SC ||
-      elementData.type === DB.ElementType.MC ||
-      elementData.type === DB.ElementType.KPRIM) &&
-    'choices' in previousResults
-  ) {
-    correctness = elementData.options.hasSampleSolution
-      ? evaluateChoicesAnswerCorrectness({ elementData, response })
-      : 1
-
-    const res = updateChoicesResults({
-      previousResults,
-      response,
-    })
-
-    return {
-      ...res,
-      correctness,
-    }
-  } else if (
-    elementData.type === DB.ElementType.NUMERICAL &&
-    'responses' in previousResults
-  ) {
-    correctness = elementData.options.hasSampleSolution
-      ? evaluateNumericalAnswerCorrectness({ elementData, response })
-      : 1
-
-    const res = updateNumericalResults({
-      previousResults,
-      elementData,
-      response,
-      correct: correctness === 1,
-    })
-
-    return {
-      ...res,
-      correctness,
-    }
-  } else if (
-    elementData.type === DB.ElementType.FREE_TEXT &&
-    'responses' in previousResults
-  ) {
-    correctness =
-      correctnessOverride ??
-      (elementData.options.hasSampleSolution
-        ? evaluateFreeTextAnswerCorrectness({ elementData, response })
-        : 1)
-
-    const res = updateFreeTextResults({
-      previousResults,
-      elementData,
-      response,
-      correct: correctness === 1,
-    })
-
-    return {
-      ...res,
-      correctness,
-    }
-  } else if (
-    elementData.type === DB.ElementType.SELECTION &&
-    'selections' in previousResults
-  ) {
-    correctness = elementData.options.hasSampleSolution
-      ? evaluateSelectionAnswerCorrectness({ elementData, response })
-      : 1
-
-    const res = updateSelectionResults({
-      previousResults,
-      response,
-    })
-
-    return {
-      ...res,
-      correctness,
-    }
-  } else if (
-    elementData.type === DB.ElementType.CASE_STUDY &&
-    'assessments' in previousResults
-  ) {
-    correctness = elementData.options.hasSampleSolution
-      ? evaluateCaseStudyAnswerCorrectness({ elementData, response })
-      : 1
-
-    const res = updateCaseStudyResults({
-      previousResults,
-      response,
-      solutions: caseStudySolutions,
-    })
-
-    return {
-      ...res,
-      correctness,
-    }
-  }
-
-  return {
-    correctness: null,
-    results: previousResults,
-    modified: true,
-  }
-}
-
-function computeAwardedPointsAndXP({
-  score,
-  xp,
-  existingResponse,
-  participation,
-  instance,
-}: {
-  score: number
-  xp: number
-  existingResponse: DB.QuestionResponse | null
-  participation: DB.Participation | null
-  instance: DB.ElementInstance
-}): {
-  pointsAwarded: number | null
-  newPointsFrom: Date | undefined
-  lastAwardedAt: Date | undefined
-  lastXpAwardedAt: Date
-  xpAwarded: number
-  newXpFrom: Date
-} {
-  const participationActive = participation?.isActive ?? false
-
-  // award points and xp based on the previous response being outside the timeframe
-  if (existingResponse) {
-    // update points (only if participation is active)
-    const pointsOutsideTimeframe =
-      !existingResponse.lastAwardedAt ||
-      dayjs(existingResponse.lastAwardedAt).isBefore(
-        dayjs().subtract(
-          instance?.options.resetTimeDays ?? POINTS_AWARD_TIMEFRAME_DAYS,
-          'days'
-        )
-      )
-
-    let pointsAwarded: number | null
-    let lastAwardedAt: Date | undefined
-    let newPointsFrom: Date | undefined
-
-    if (participationActive) {
-      pointsAwarded = pointsOutsideTimeframe ? score : 0
-      lastAwardedAt =
-        pointsOutsideTimeframe || !existingResponse.lastAwardedAt
-          ? new Date()
-          : existingResponse.lastAwardedAt
-      newPointsFrom = dayjs(lastAwardedAt)
-        .add(
-          instance?.options.resetTimeDays ?? POINTS_AWARD_TIMEFRAME_DAYS,
-          'days'
-        )
-        .toDate()
-    } else {
-      pointsAwarded = null
-      lastAwardedAt = undefined
-      newPointsFrom = undefined
-    }
-
-    // update experience points (independent of active participation -> stored directly on participant)
-    const xpOutsideTimeframe =
-      !existingResponse.lastXpAwardedAt ||
-      dayjs(existingResponse.lastXpAwardedAt).isBefore(
-        dayjs().subtract(XP_AWARD_TIMEFRAME_DAYS, 'days')
-      )
-
-    const xpAwarded = xpOutsideTimeframe ? xp : 0
-    const lastXpAwardedAt =
-      xpOutsideTimeframe || !existingResponse.lastXpAwardedAt
-        ? new Date()
-        : existingResponse.lastXpAwardedAt
-    const newXpFrom = dayjs(lastXpAwardedAt)
-      .add(XP_AWARD_TIMEFRAME_DAYS, 'days')
-      .toDate()
-
-    return {
-      pointsAwarded,
-      newPointsFrom,
-      lastAwardedAt,
-      xpAwarded,
-      newXpFrom,
-      lastXpAwardedAt,
-    }
-  }
-
-  // no previous response exists -> award points and xp based on the current response and if participation is active
-  const lastAwardedAt = participationActive ? new Date() : undefined
-  const newPointsFrom = participationActive
-    ? dayjs(lastAwardedAt)
-        .add(
-          instance?.options.resetTimeDays ?? POINTS_AWARD_TIMEFRAME_DAYS,
-          'days'
-        )
-        .toDate()
-    : undefined
-  const newXpFrom = dayjs(lastAwardedAt)
-    .add(XP_AWARD_TIMEFRAME_DAYS, 'days')
-    .toDate()
-
-  return {
-    pointsAwarded: participationActive ? score : null,
-    newPointsFrom,
-    lastAwardedAt,
-    xpAwarded: xp,
-    newXpFrom,
-    lastXpAwardedAt: new Date(),
-  }
-}
-
-function computeAggregatedResponsesChoices({
-  instance,
-  existingResponse,
-  response,
-}: {
-  instance: DB.ElementInstance
-  existingResponse: DB.QuestionResponse | null
-  response: ResponseInput
-}): ElementResultsChoices {
-  const newAggResponses = (existingResponse?.aggregatedResponses ??
-    getInitialInstanceResults(instance.elementData)) as ElementResultsChoices
-
-  // update aggregated responses for choices
-  newAggResponses.choices = (
-    response as SingleQuestionResponseChoices
-  ).choices.reduce((acc, choiceResponse) => {
-    acc[choiceResponse.ix] = (acc[choiceResponse.ix] ?? 0) + 1
-    return acc
-  }, newAggResponses.choices)
-  newAggResponses.total = newAggResponses.total + 1
-
-  return newAggResponses
-}
-
-function computeAggregatedResponsesOpen({
-  instance,
-  existingResponse,
-  responseValue,
-  correctness,
-}: {
-  instance: DB.ElementInstance
-  existingResponse: DB.QuestionResponse | null
-  responseValue: string
-  correctness: number
-}) {
-  const newAggResponses = (existingResponse?.aggregatedResponses ??
-    getInitialInstanceResults(instance.elementData)) as ElementResultsOpen
-
-  // update aggregated responses for open questions
-  const MD5 = createHash('md5')
-  MD5.update(responseValue)
-  const hashedValue = MD5.digest('hex')
-
-  if (Object.keys(newAggResponses.responses).includes(hashedValue)) {
-    newAggResponses.responses = {
-      ...newAggResponses.responses,
-      [hashedValue]: {
-        ...newAggResponses.responses[hashedValue]!,
-        count: newAggResponses.responses[hashedValue]!.count + 1,
-      },
-    }
-  } else {
-    newAggResponses.responses = {
-      ...newAggResponses.responses,
-      [hashedValue]: {
-        value: responseValue,
-        count: 1,
-        correct: correctness === 1,
-      },
-    }
-  }
-  newAggResponses.total = newAggResponses.total + 1
-
-  return newAggResponses
-}
-
-function computeAggregatedResponsesSelection({
-  instance,
-  existingResponse,
-  responseSelection,
-}: {
-  instance: DB.ElementInstance
-  existingResponse: DB.QuestionResponse | null
-  responseSelection: number[]
-}) {
-  if (!('answerCollection' in instance.elementData.options)) {
-    throw new Error('Answer collection is missing in selection element')
-  }
-
-  const newAggResponses = (existingResponse?.aggregatedResponses ??
-    getInitialInstanceResults(instance.elementData)) as ElementResultsSelection
-
-  // increment all entries that are in response selection
-  const updatedSelections = { ...newAggResponses.selections }
-  responseSelection.forEach((ix) => {
-    if (ix in updatedSelections && typeof updatedSelections[ix] === 'number') {
-      updatedSelections[ix] = updatedSelections[ix] + 1
-    }
-  })
-
-  newAggResponses.selections = updatedSelections
-  newAggResponses.total = newAggResponses.total + 1
-  return newAggResponses
-}
-
-function computeAggregatedResponsesCaseStudy({
-  instance,
-  existingResponse,
-  responseAssessment,
-  solutions,
-}: {
-  instance: DB.ElementInstance
-  existingResponse: DB.QuestionResponse | null
-  responseAssessment: CaseStudyCaseResponse[]
-  solutions?: CaseStudySolutionsObject
-}) {
-  // check that all data required for results initialization is provided in instance
-  if (
-    !('items' in instance.elementData.options) ||
-    !('cases' in instance.elementData.options) ||
-    !('criteria' in instance.elementData.options)
-  ) {
-    throw new Error(
-      'Items, cases, or criteria are missing in case study element'
-    )
-  }
-
-  // initialize aggregated responses either empty or with previous values
-  const newAggResponses = (existingResponse?.aggregatedResponses ??
-    getInitialInstanceResults(instance.elementData)) as ElementResultsCaseStudy
-
-  const updatedResults = updateCaseStudyResults({
-    previousResults: newAggResponses,
-    response: { assessment: responseAssessment },
-    solutions,
-  })
-
-  return updatedResults.results
-}
-
-function computeAggregatedResponsesQuestion({
-  instance,
-  existingResponse,
-  response,
-  correctness,
-  caseStudySolutions,
-}: {
-  instance: DB.ElementInstance
-  existingResponse: DB.QuestionResponse | null
-  response: ResponseInput
-  correctness?: number | null
-  caseStudySolutions?: CaseStudySolutionsObject
-}): ElementInstanceResults | null {
-  if (
-    instance.elementType === DB.ElementType.SC ||
-    instance.elementType === DB.ElementType.MC ||
-    instance.elementType === DB.ElementType.KPRIM
-  ) {
-    return computeAggregatedResponsesChoices({
-      instance,
-      existingResponse,
-      response,
-    })
-  } else if (
-    instance.elementType === DB.ElementType.NUMERICAL ||
-    instance.elementType === DB.ElementType.FREE_TEXT
-  ) {
-    return computeAggregatedResponsesOpen({
-      instance,
-      existingResponse,
-      responseValue:
-        instance.elementType === DB.ElementType.NUMERICAL
-          ? String(parseFloat(response.value!))
-          : toLowerCase(response.value!.trim()),
-      correctness: correctness ?? 0,
-    })
-  } else if (instance.elementType === DB.ElementType.SELECTION) {
-    return computeAggregatedResponsesSelection({
-      instance,
-      existingResponse,
-      responseSelection: response.selection!,
-    })
-  } else if (instance.elementType === DB.ElementType.CASE_STUDY) {
-    return computeAggregatedResponsesCaseStudy({
-      instance,
-      existingResponse,
-      responseAssessment: response.assessment!,
-      solutions: caseStudySolutions,
-    })
-  }
-
-  return null
-}
-
-async function upsertQuestionResponse({
-  prisma,
-  id,
-  participantId,
-  courseId,
-  response,
-  correctness,
-  score,
-  pointsAwarded,
-  lastAwardedAt,
-  xpAwarded,
-  lastXpAwardedAt,
-  newAverageResponseTime,
-  existingResponse,
-  newAggResponses,
-  practiceQuizId,
-  microLearningId,
-  resultSpacedRepetition,
-}: {
-  prisma: PrismaTransactionClient
-  id: number
-  participantId: string
-  courseId: string
-  response: ResponseInput
-  correctness: number
-  score: number
-  pointsAwarded: number | null
-  lastAwardedAt: Date
-  xpAwarded: number
-  lastXpAwardedAt: Date
-  newAverageResponseTime: number
-  existingResponse: DB.QuestionResponse | null
-  newAggResponses: ElementInstanceResults
-  practiceQuizId?: string
-  microLearningId?: string
-  resultSpacedRepetition: SpacedRepetitionResult
-}) {
-  const responseCorrectness =
-    correctness === 1
-      ? DB.ResponseCorrectness.CORRECT
-      : correctness === 0
-        ? DB.ResponseCorrectness.WRONG
-        : DB.ResponseCorrectness.PARTIAL
-
-  await prisma.questionResponse.upsert({
-    where: {
-      participantId_elementInstanceId: {
-        participantId,
-        elementInstanceId: id,
-      },
-    },
-    create: {
-      totalScore: score,
-      totalPointsAwarded: pointsAwarded,
-      totalXpAwarded: xpAwarded,
-      trialsCount: 1,
-      averageTimeSpent: newAverageResponseTime ?? 0,
-      lastAwardedAt,
-      lastXpAwardedAt,
-      firstResponse: response as SingleQuestionResponse,
-      firstResponseCorrectness: responseCorrectness,
-      lastResponse: response as SingleQuestionResponse,
-      lastResponseCorrectness: responseCorrectness,
-      aggregatedResponses: newAggResponses,
-      participant: {
-        connect: { id: participantId },
-      },
-      elementInstance: {
-        connect: { id },
-      },
-      practiceQuiz: practiceQuizId
-        ? { connect: { id: practiceQuizId } }
-        : undefined,
-      microLearning: microLearningId
-        ? { connect: { id: microLearningId } }
-        : undefined,
-      course: {
-        connect: {
-          id: courseId,
-        },
-      },
-      participation: {
-        connect: {
-          courseId_participantId: {
-            courseId,
-            participantId,
-          },
-        },
-      },
-
-      // compute and store new correctness parameters
-      ...combineNewCorrectnessParams({
-        correct: correctness === 1,
-        partial: correctness > 0 && correctness < 1,
-        incorrect: correctness === 0,
-      }),
-
-      eFactor: resultSpacedRepetition.efactor,
-      nextDueAt: resultSpacedRepetition.nextDueAt,
-      interval: resultSpacedRepetition.interval,
-    },
-    update: {
-      lastResponse: response as SingleQuestionResponse,
-      lastResponseCorrectness: responseCorrectness,
-      aggregatedResponses: newAggResponses,
-      lastAwardedAt,
-      lastXpAwardedAt,
-      trialsCount: {
-        increment: 1,
-      },
-      averageTimeSpent: newAverageResponseTime ?? 0,
-      totalScore: {
-        increment: score,
-      },
-      totalPointsAwarded:
-        typeof pointsAwarded === 'number' ? { increment: pointsAwarded } : null,
-      totalXpAwarded: {
-        increment: xpAwarded,
-      },
-
-      ...combineCorrectnessParams({
-        correct: correctness === 1,
-        partial: correctness > 0 && correctness < 1,
-        incorrect: correctness === 0,
-        existingResponse,
-      }),
-
-      eFactor: resultSpacedRepetition.efactor,
-      nextDueAt: resultSpacedRepetition.nextDueAt,
-      interval: resultSpacedRepetition.interval,
-    },
-  })
-}
-
-async function createQuestionResponseDetail({
-  prisma,
-  id,
-  participantId,
-  courseId,
-  response,
-  score,
-  pointsAwarded,
-  xpAwarded,
-  answerTime,
-  practiceQuizId,
-  microLearningId,
-}: {
-  prisma: PrismaTransactionClient
-  id: number
-  participantId: string
-  courseId: string
-  response: ResponseInput
-  score: number
-  pointsAwarded: number | null
-  xpAwarded: number
-  answerTime: number
-  practiceQuizId?: string
-  microLearningId?: string
-}) {
-  return await prisma.questionResponseDetail.create({
-    data: {
-      score,
-      pointsAwarded,
-      xpAwarded,
-      timeSpent: answerTime,
-      response: response as SingleQuestionResponse,
-      participant: {
-        connect: { id: participantId },
-      },
-      elementInstance: {
-        connect: { id },
-      },
-      practiceQuiz: practiceQuizId
-        ? { connect: { id: practiceQuizId } }
-        : undefined,
-      microLearning: microLearningId
-        ? { connect: { id: microLearningId } }
-        : undefined,
-      participation: {
-        connect: {
-          courseId_participantId: {
-            courseId,
-            participantId,
-          },
-        },
-      },
-    },
-  })
-}
-
-async function incrementParticipantXp({
-  prisma,
-  participantId,
-  xpAwarded,
-}: {
-  prisma: PrismaTransactionClient
-  participantId: string
-  xpAwarded: number
-}) {
-  await prisma.participant.update({
-    where: {
-      id: participantId,
-    },
-    data: {
-      xp: {
-        increment: xpAwarded,
-      },
-    },
-  })
-}
-
-async function updateLeaderboardOnQuestionResponse({
-  prisma,
-  participantId,
-  courseId,
-  pointsAwarded,
-}: {
-  prisma: PrismaTransactionClient
-  participantId: string
-  courseId: string
-  pointsAwarded: number
-}) {
-  await prisma.leaderboardEntry.upsert({
-    where: {
-      type_participantId_courseId: {
-        type: 'COURSE',
-        courseId,
-        participantId,
-      },
-    },
-    create: {
-      type: 'COURSE',
-      score: pointsAwarded,
-      participant: {
-        connect: {
-          id: participantId,
-        },
-      },
-      course: {
-        connect: {
-          id: courseId,
-        },
-      },
-      participation: {
-        connect: {
-          courseId_participantId: {
-            courseId,
-            participantId,
-          },
-        },
-      },
-    },
-    update: {
-      score: {
-        increment: pointsAwarded,
-      },
-    },
-  })
-}
-
 type RespondToQuestionArgs = {
   id: number
   courseId: string
@@ -2617,19 +923,13 @@ type RespondToQuestionArgs = {
   participation: (DB.Participation & { participant: DB.Participant }) | null
   skipTracking?: boolean
   correctnessOverride?: number
-  freeTextAttemptId?: string
+  awardOverride?: {
+    pointsAwarded: number | null
+    xpAwarded: number
+  }
 }
 
 export async function respondToQuestion(
-  args: RespondToQuestionArgs,
-  ctx: Context
-) {
-  return await ctx.prisma.$transaction(async (prisma) =>
-    respondToQuestionInTransaction(args, ctx, prisma)
-  )
-}
-
-export async function respondToQuestionInTransaction(
   {
     id,
     courseId,
@@ -2638,362 +938,32 @@ export async function respondToQuestionInTransaction(
     participation,
     skipTracking,
     correctnessOverride,
-    freeTextAttemptId,
+    awardOverride,
   }: RespondToQuestionArgs,
-  ctx: Context,
-  prisma: PrismaTransactionClient
+  ctx: Context
 ) {
-  const freeTextAttempt = freeTextAttemptId
-    ? await prisma.freeTextAttempt.findUnique({
-        where: { id: freeTextAttemptId },
-        include: { cycle: true },
-      })
-    : null
-  if (
-    freeTextAttemptId &&
-    (!freeTextAttempt ||
-      freeTextAttempt.questionResponseDetailId !== null ||
-      freeTextAttempt.evaluationStatus !==
-        DB.FreeTextEvaluationStatus.EVALUATED)
-  ) {
-    return null
-  }
-
-  const existingInstance = await getValidateElementInstance({
-    prisma,
-    id,
-    participantId:
-      ctx.user?.role === DB.UserRole.PARTICIPANT ? ctx.user?.sub : undefined,
-  })
-
-  // if the instance does not exist or the elementData is not defined, return early
-  if (!existingInstance || !existingInstance?.elementData) {
-    return null
-  }
-
-  const existingResponse =
-    existingInstance.responses &&
-    existingInstance.responses.length > 0 &&
-    existingInstance.responses[0]
-      ? existingInstance.responses[0]
+  const actor =
+    ctx.user?.role === DB.UserRole.PARTICIPANT && participation
+      ? { participation }
       : null
 
-  // conver the case study solutions to object form for more efficient access
-  const caseStudySolutions =
-    existingInstance.elementType === DB.ElementType.CASE_STUDY
-      ? convertCaseStudySolutionsObject({
-          instance: existingInstance,
-        })
-      : undefined
-
-  // evaluate response correctness and compute updated instance results
-  const { correctness, results, modified } = updateQuestionResults({
-    existingInstance,
-    participation,
-    response,
-    caseStudySolutions,
-    correctnessOverride,
-  })
-
-  if (!modified || results === null) {
-    return null
-  }
-
-  // average answer time computations if participant is logged in
-  const { newAverageResponseTime, newAverageInstanceTime } = participation
-    ? computeNewAverageTimes({
-        existingInstance,
-        existingResponse,
+  return await ctx.prisma.$transaction((prisma) =>
+    applyQuestionResponseInTransaction(
+      {
+        id,
+        courseId,
+        response,
         answerTime,
-      })
-    : {
-        newAverageInstanceTime: undefined,
-        newAverageResponseTime: answerTime,
-      }
-
-  // compute updated instance statistics
-  const instanceInPracticeQuiz = !!existingInstance.elementStack?.practiceQuizId
-  const statisticsUpdate = computeUpdatedInstanceStatistics({
-    participation,
-    existingResponse: existingResponse, // this is safe because we only use it inside the function if the participation is defined
-    newAverageInstanceTime,
-    answerCorrect: correctness === 1,
-    answerPartial: (correctness ?? 0) < 1 && (correctness ?? 0) > 0,
-    answerIncorrect: correctness === 0,
-    instanceInPracticeQuiz,
-  })
-
-  // update the element instance
-  const updatedInstance = !skipTracking
-    ? await prisma.elementInstance.update({
-        where: { id },
-        data: {
-          results: participation ? results : undefined,
-          anonymousResults: participation ? undefined : results,
-          instanceStatistics: statisticsUpdate,
-        },
-        include: {
-          elementStack: true,
-        },
-      })
-    : existingInstance
-
-  // compute question evaluation
-  const questionEval = computeQuestionEvaluation({
-    elementData: existingInstance.elementData,
-    results: updatedInstance.results,
-    anonymousResults: updatedInstance.anonymousResults,
-    correctness,
-    multiplier: updatedInstance.options.pointsMultiplier,
-  })
-
-  // processing of percentile into status of instance
-  const percentile = questionEval?.percentile ?? 0
-  const status =
-    percentile === 0
-      ? StackFeedbackStatus.INCORRECT
-      : percentile === 1
-        ? StackFeedbackStatus.CORRECT
-        : StackFeedbackStatus.PARTIAL
-
-  // if participant is not logged in, return early and return the evaluation
-  if (
-    !questionEval ||
-    !participation ||
-    !ctx.user?.sub ||
-    ctx.user.role !== DB.UserRole.PARTICIPANT
-  ) {
-    return {
-      ...updatedInstance,
-      evaluation: questionEval
-        ? {
-            ...questionEval,
-            pointsAwarded: undefined,
-            newPointsFrom: undefined,
-            xpAwarded: undefined,
-            newXpFrom: undefined,
-          }
-        : undefined,
-      status,
-    }
-  }
-
-  // compute the awarded points & XP and all associated dates
-  let {
-    pointsAwarded,
-    newPointsFrom,
-    lastAwardedAt,
-    lastXpAwardedAt,
-    xpAwarded,
-    newXpFrom,
-  } = computeAwardedPointsAndXP({
-    score: questionEval.score ?? 0,
-    xp: questionEval.xp ?? 0,
-    existingResponse,
-    participation,
-    instance: updatedInstance,
-  })
-
-  if (freeTextAttempt) {
-    const previousBest = await prisma.freeTextAttempt.aggregate({
-      where: {
-        cycleId: freeTextAttempt.cycleId,
-        id: { not: freeTextAttempt.id },
-        evaluationStatus: DB.FreeTextEvaluationStatus.EVALUATED,
+        actor,
+        skipTracking,
+        correctnessOverride,
+        awardOverride,
       },
-      _max: { aggregateScore: true },
-    })
-    const previousPercentage = (previousBest._max.aggregateScore ?? 0) / 100
-    const previousScore =
-      previousPercentage *
-      POINTS_PER_INSTANCE *
-      (updatedInstance.options.pointsMultiplier ?? 1)
-    const previousXp = computeAwardedXp({
-      pointsPercentage: previousPercentage,
-    })
-    pointsAwarded = participation.isActive
-      ? freeTextAttempt.cycle.pointsRewardEligible
-        ? Math.max(0, (questionEval.score ?? 0) - previousScore)
-        : 0
-      : null
-    xpAwarded = freeTextAttempt.cycle.xpRewardEligible
-      ? Math.max(0, (questionEval.xp ?? 0) - previousXp)
-      : 0
-    if ((pointsAwarded ?? 0) > 0) {
-      lastAwardedAt = new Date()
-      newPointsFrom = dayjs(lastAwardedAt)
-        .add(
-          updatedInstance.options.resetTimeDays ?? POINTS_AWARD_TIMEFRAME_DAYS,
-          'days'
-        )
-        .toDate()
-    }
-    if (xpAwarded > 0) {
-      lastXpAwardedAt = new Date()
-      newXpFrom = dayjs(lastXpAwardedAt)
-        .add(XP_AWARD_TIMEFRAME_DAYS, 'days')
-        .toDate()
-    }
-  }
-
-  // create a question response detail
-  if (!skipTracking) {
-    // compute updated aggregated responses
-    const newAggResponses = computeAggregatedResponsesQuestion({
-      instance: updatedInstance,
-      existingResponse,
-      response,
-      correctness,
-      caseStudySolutions,
-    })
-
-    if (!newAggResponses) {
-      throw new Error(
-        `Failed to compute aggregated responses for question type ${updatedInstance.elementType}`
-      )
-    }
-
-    // update aggregated results for choices and open questions
-    const streakIncrement = percentile === 1 ? 1 : 0
-    const resultSpacedRepetition = updateSpacedRepetition({
-      eFactor: existingResponse?.eFactor ?? 2.5,
-      interval: existingResponse?.interval ?? 1,
-      streak: (existingResponse?.correctCountStreak ?? 0) + streakIncrement,
-      grade: percentile,
-    })
-
-    const responseDetail = await createQuestionResponseDetail({
-      prisma,
-      id,
-      participantId: ctx.user.sub,
-      courseId,
-      response,
-      score: questionEval.score ?? 0,
-      pointsAwarded,
-      xpAwarded,
-      answerTime,
-      practiceQuizId: updatedInstance.elementStack?.practiceQuizId ?? undefined,
-      microLearningId:
-        updatedInstance.elementStack?.microLearningId ?? undefined,
-    })
-
-    if (freeTextAttempt) {
-      await prisma.freeTextAttempt.update({
-        where: { id: freeTextAttempt.id },
-        data: { questionResponseDetailId: responseDetail.id },
-      })
-      await prisma.freeTextPracticeCycle.update({
-        where: { id: freeTextAttempt.cycleId },
-        data: {
-          pointsAwarded: { increment: pointsAwarded ?? 0 },
-          xpAwarded: { increment: xpAwarded },
-          bestXp: Math.max(freeTextAttempt.cycle.bestXp, questionEval.xp ?? 0),
-        },
-      })
-    }
-
-    // upsert the question response
-    await upsertQuestionResponse({
-      prisma,
-      id,
-      participantId: ctx.user.sub,
-      courseId,
-      response,
-      correctness: percentile,
-      score: questionEval.score ?? 0,
-      pointsAwarded,
-      lastAwardedAt: lastAwardedAt ?? new Date(),
-      xpAwarded,
-      lastXpAwardedAt: lastXpAwardedAt ?? new Date(),
-      newAverageResponseTime,
-      existingResponse,
-      newAggResponses,
-      practiceQuizId: updatedInstance.elementStack?.practiceQuizId ?? undefined,
-      microLearningId:
-        updatedInstance.elementStack?.microLearningId ?? undefined,
-      resultSpacedRepetition,
-    })
-
-    // increment participant xp
-    if (xpAwarded > 0) {
-      await incrementParticipantXp({
-        prisma,
-        participantId: ctx.user.sub,
-        xpAwarded,
-      })
-    }
-
-    // create or increment the leaderboard entry, if the participant has an active participation in the course
-    // active participation has already been checked during computation of pointsAwarded
-    if (typeof pointsAwarded === 'number' && pointsAwarded !== null) {
-      await updateLeaderboardOnQuestionResponse({
-        prisma,
-        participantId: ctx.user.sub,
-        courseId,
-        pointsAwarded,
-      })
-    }
-
-    // if either XP or points are awarded, update the daily student timeline entry
-    if (
-      xpAwarded > 0 ||
-      (typeof pointsAwarded === 'number' && pointsAwarded !== null)
-    ) {
-      await upsertDailyTimelineEntry({
-        prisma,
-        participantId: ctx.user.sub,
-        courseId,
-        xpAwarded,
-        pointsAwarded: pointsAwarded ?? undefined,
-      })
-    }
-  }
-
-  return {
-    ...updatedInstance,
-    evaluation: {
-      ...questionEval,
-      pointsAwarded,
-      newPointsFrom,
-      xpAwarded,
-      newXpFrom,
-    },
-    status,
-  }
-}
-
-export async function applyFreeTextAttemptResponse(
-  args: { attemptId: string },
-  prisma: DB.PrismaClient
-) {
-  return await prisma.$transaction(async (tx) =>
-    applyFreeTextAttemptResponseInTransaction(args, tx)
+      prisma
+    )
   )
 }
 
-export async function applyFreeTextAttemptResponseInTransaction(
-  { attemptId }: { attemptId: string },
-  prisma: PrismaTransactionClient
-) {
-  return await applyFreeTextAttemptResponseCore(
-    { attemptId },
-    prisma,
-    respondToQuestionInTransaction
-  )
-}
-
-export async function createAndApplyFreeTextAttempt(
-  input: CreateFreeTextAttemptInput,
-  ctx: ContextWithUser,
-  options?: FreeTextEvaluationServiceOptions
-) {
-  return await createFreeTextAttempt(
-    input,
-    ctx,
-    applyFreeTextAttemptResponseInTransaction,
-    options
-  )
-}
 // #endregion
 
 // ! Element & Stack Response & Combination Logic
@@ -3198,50 +1168,13 @@ async function respondToElement({
             'Semantic free-text responses require an answer and client submission ID'
           )
         }
-        const semanticState = await createAndApplyFreeTextAttempt(
-          {
-            instanceId: response.instanceId,
-            answer: response.freeTextResponse,
-            answerTime,
-            clientSubmissionId: response.clientSubmissionId,
-          },
-          ctx as ContextWithUser
-        )
-        const currentAttempt = semanticState.currentAttempt
-        const grading =
-          currentAttempt?.correctness === DB.FreeTextCorrectnessCategory.CORRECT
-            ? StackFeedbackStatus.CORRECT
-            : currentAttempt?.correctness ===
-                DB.FreeTextCorrectnessCategory.PARTIAL
-              ? StackFeedbackStatus.PARTIAL
-              : currentAttempt?.correctness ===
-                  DB.FreeTextCorrectnessCategory.INCORRECT
-                ? StackFeedbackStatus.INCORRECT
-                : null
-        const score =
-          typeof currentAttempt?.aggregateScore === 'number'
-            ? (currentAttempt.aggregateScore / 100) *
-              POINTS_PER_INSTANCE *
-              (instance?.options.pointsMultiplier ?? 1)
-            : null
-
-        return {
-          grading,
-          score,
-          evaluation: {
-            instanceId: response.instanceId,
-            elementType: DB.ElementType.FREE_TEXT,
-            score: score ?? 0,
-            pointsMultiplier: instance?.options.pointsMultiplier ?? 1,
-            explanation: semanticState.solutionAuthorized
-              ? semanticState.explanation
-              : null,
-            feedbacks: [],
-            answers: [],
-            solutions: [],
-            semanticState,
-          } as InstanceEvaluationFreeText,
-        }
+        if (!instance) return { grading: null, score: null, evaluation: null }
+        return await submitSemanticFreeTextPracticeResponse({
+          ctx: ctx as ContextWithUser,
+          response,
+          answerTime,
+          instance,
+        })
       }
     }
 
@@ -3449,246 +1382,6 @@ type CommonEvaluationProps = {
   hasAnswerFeedbacks: boolean
 }
 
-function combineChoicesResults({
-  choices,
-  results,
-  anonymousResults,
-}: {
-  choices: Choice[]
-  results: ElementResultsChoices['choices']
-  anonymousResults: ElementResultsChoices['choices']
-}) {
-  return choices.map((choice) => ({
-    ix: choice.ix,
-    value: choice.value,
-    count: (results[choice.ix] ?? 0) + (anonymousResults[choice.ix] ?? 0),
-    correct: choice.correct,
-    feedback: choice.feedback,
-  }))
-}
-
-function combineNumericalResults({
-  results,
-  anonymousResults,
-}: {
-  results: ElementResultsOpen
-  anonymousResults: ElementResultsOpen
-}) {
-  return [
-    ...Object.values(results.responses),
-    ...Object.values(anonymousResults.responses),
-  ].reduce<{ value: number; count: number; correct?: boolean | null }[]>(
-    (acc, response) => {
-      const responseValue = parseFloat(response.value)
-      const ix = acc.findIndex(
-        (r) => Math.abs(r.value - responseValue) < Number.EPSILON
-      )
-      if (ix === -1) {
-        acc.push({
-          value: responseValue,
-          count: response.count,
-          correct: response.correct,
-        })
-      } else {
-        acc[ix] = {
-          ...acc[ix]!,
-          count: acc[ix]!.count + response.count,
-        }
-      }
-      return acc
-    },
-    []
-  )
-}
-
-function combineFreeTextResults({
-  results,
-  anonymousResults,
-}: {
-  results: ElementResultsOpen
-  anonymousResults: ElementResultsOpen
-}) {
-  return [
-    ...Object.values(results.responses),
-    ...Object.values(anonymousResults.responses),
-  ].reduce<{ value: string; count: number; correct?: boolean | null }[]>(
-    (acc, response) => {
-      const ix = acc.findIndex((r) => r.value === response.value)
-      if (ix === -1) {
-        acc.push({
-          value: response.value,
-          count: response.count,
-          correct: response.correct,
-        })
-      } else {
-        acc[ix] = {
-          ...acc[ix]!,
-          count: acc[ix]!.count + response.count,
-        }
-      }
-      return acc
-    },
-    []
-  )
-}
-
-function combineSelectionResults({
-  results,
-  anonymousResults,
-  answerOptions,
-}: {
-  results: ElementResultsSelection
-  anonymousResults: ElementResultsSelection
-  answerOptions: ElementOptionsAnswerCollection
-}) {
-  return answerOptions.entries.map((option) => ({
-    answerId: option.id,
-    value: option.value,
-    count:
-      (results.selections[option.id] ?? 0) +
-      (anonymousResults.selections[option.id] ?? 0),
-  }))
-}
-
-function reduceCaseStudyResults({
-  results,
-  anonymousResults,
-  options,
-}: {
-  results: ElementResultsCaseStudy
-  anonymousResults: ElementResultsCaseStudy
-  options: ElementOptionsCaseStudy
-}): SingleCaseStudyResponse[] {
-  const combinedResponses = options.cases.flatMap((caseObj) =>
-    options.items
-      ? options.items.flatMap((item) =>
-          options.criteria.map((criterion) => {
-            const caseId = caseObj.id
-            const itemId = item.id
-            const criterionId = criterion.id
-
-            // extract result and anonymous result values
-            const resultValues = Object.values(
-              results.assessments[caseId]?.[itemId]?.[criterionId] ?? {}
-            ).map((r) => r.value)
-            const anonymousResultValues = Object.values(
-              anonymousResults.assessments[caseId]?.[itemId]?.[criterionId] ??
-                {}
-            ).map((r) => r.value)
-
-            // combine the values into a single array with unique values
-            const combinedResultValues = [
-              ...new Set([...resultValues, ...anonymousResultValues]),
-            ]
-
-            return {
-              caseId,
-              itemId,
-              criterionId,
-              responseValues: combinedResultValues,
-            }
-          })
-        )
-      : []
-  )
-
-  return combinedResponses
-}
-
-function combineCaseStudyResults({
-  results,
-  anonymousResults,
-  options,
-}: {
-  results: ElementResultsCaseStudy
-  anonymousResults: ElementResultsCaseStudy
-  options: ElementOptionsCaseStudy
-}): ICaseStudyElementEvaluationResults['caseResults'] {
-  return options.cases.map((caseObj) => {
-    const caseSolutions = caseObj.solutions
-
-    return {
-      caseId: caseObj.id,
-      items:
-        options.items?.map((item) => {
-          const itemSolutions = caseSolutions?.find((s) => s.itemId === item.id)
-
-          return {
-            itemId: item.id,
-            criteria: options.criteria.map((criterion) => {
-              const criterionSolution = itemSolutions?.criteriaSolutions.find(
-                (c) => c.criterionId === criterion.id
-              )
-              const criterionResults =
-                results.assessments[caseObj.id]?.[item.id]?.[criterion.id]
-              const criterionAnonymousResults =
-                anonymousResults.assessments[caseObj.id]?.[item.id]?.[
-                  criterion.id
-                ]
-
-              // merge the results and anonymous results into a single object
-              const mergedResults = [
-                ...Object.entries(criterionResults ?? {}),
-                ...Object.entries(criterionAnonymousResults ?? {}),
-              ].reduce<{
-                [valueHash: string]: {
-                  value: number
-                  count: number
-                  correct?: boolean
-                }
-              }>((acc, [key, entry]) => {
-                // if the key already exists in acc, sum up the counts
-                if (acc[key]) {
-                  acc[key] = {
-                    value: acc[key]!.value,
-                    count: acc[key]!.count + entry.count,
-                    correct: acc[key]!.correct ?? entry.correct,
-                  }
-                } else {
-                  acc[key] = entry
-                }
-                return acc
-              }, {})
-
-              // extract values from merged results
-              const responses = Object.values(mergedResults)
-
-              // if the criterion is a liker criterion, make sure the entire valid bar is included in the solution interval
-              const isLikertCriterion = !!criterion.labels
-
-              return {
-                criterionId: criterion.id,
-                name: criterion.name,
-                min: criterion.min,
-                max: criterion.max,
-                step: criterion.step,
-                unit: criterion.unit,
-                labels: criterion.labels,
-
-                solutionMin: criterionSolution?.min
-                  ? isLikertCriterion
-                    ? criterionSolution?.min - 0.5
-                    : criterionSolution?.min
-                  : undefined,
-                solutionMax: criterionSolution?.max
-                  ? isLikertCriterion
-                    ? criterionSolution?.max + 0.5
-                    : criterionSolution?.max
-                  : undefined,
-
-                statistics:
-                  responses && responses.length > 0
-                    ? (computeNumericalStatistics(responses) ?? undefined)
-                    : undefined,
-                responses,
-              }
-            }),
-          }
-        }) ?? [],
-    }
-  })
-}
-
 function computeChoicesEvaluation({
   options,
   results,
@@ -3715,31 +1408,6 @@ function computeChoicesEvaluation({
       choices,
     },
   }
-}
-
-function computeNumericalStatistics(
-  results: {
-    value: number
-    count: number
-    correct?: boolean | null
-  }[]
-) {
-  const valueArray = results.reduce<number[]>((acc, { count, value }) => {
-    const elements = Array(count).fill(value)
-    return acc.concat(elements)
-  }, [])
-
-  return valueArray.length > 0
-    ? {
-        max: max(valueArray),
-        mean: mean(valueArray),
-        median: median(valueArray),
-        min: min(valueArray),
-        q1: quantileSeq(valueArray, 0.25) as number,
-        q3: quantileSeq(valueArray, 0.75) as number,
-        sd: std(valueArray) as unknown as number, // since arrays are guaranteed to be flat -> return type always number
-      }
-    : null
 }
 
 function computeNumericalEvaluation({
